#2 | Alvaro | added persistency in database

+++ src/main/kotlin/com/wefox/kanekotic/centralizedPayments/models/Payment.kt
@@ -1,4 +1,4 @@
 package com.wefox.kanekotic.centralizedPayments.models
 
-data class Payment(val payment_id: String, val account_id: Int, val payment_type: String, val credit_card: String, val amount: Int, val delay: Int)
+data class Payment(val payment_id: String, val account_id: Int, val payment_type: String, val credit_card: String, val amount: Double, val delay: Int)
